mte-main/requirements.txt
numpy=1.23.1
mindspore-gpu=1.10.1
easydict
pyyaml
onnxruntime-gpu
morfessor
pyicu
pycld2
pyenchant
polyglot